requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
diff-match-patch==20241021
Django==4.2.16
django-appconf==1.0.6
django-cryptography-django5==2.2
django-grappelli==4.0.1
django-import-export==4.3.1
django-rangefilter==0.9.1
django-sql-explorer==5.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-http-client==3.3.7
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tablib==3.7.0
typing_extensions==4.12.2
urllib3==2.2.3
